refactor: move icons.ts to shared/icons.ts
