+++ packages/base/src/index.ts
@@ -2,7 +2,7 @@ export * from './features/annotations';
 export * from './commands/index';
 export * from './constants';
 export * from './dialogs/layerCreationFormDialog';
-export * from './icons';
+export * from './shared/icons';
 export * from './mainview';
 export * from './menus';
 export * from './workspace/panels';
