+++ packages/base/src/shared/icons.ts
@@ -7,25 +7,25 @@
 
 import { LabIcon } from '@jupyterlab/ui-components';
 
-import bookOpenSvgStr from '../style/icons/book_open.svg';
-import clockSvgStr from '../style/icons/clock-solid.svg';
-import geoJsonSvgStr from '../style/icons/geojson.svg';
-import geolocationSvgStr from '../style/icons/geolocation.svg';
-import geoPackageSvgStr from '../style/icons/geopackage.svg';
-import infoSvgStr from '../style/icons/info-solid.svg';
-import logoSvgStr from '../style/icons/logo.svg';
-import logoMiniSvgStr from '../style/icons/logo_mini.svg';
-import logoMiniAlternativeSvgStr from '../style/icons/logo_mini_alternative.svg';
-import logoMiniQGZ from '../style/icons/logo_mini_qgz.svg';
-import markerSvgStr from '../style/icons/marker.svg';
-import moundSvgStr from '../style/icons/mound.svg';
-import nonVisibilitySvgStr from '../style/icons/nonvisibility.svg';
-import rasterSvgStr from '../style/icons/raster.svg';
-import targetWithCenterSvgStr from '../style/icons/target_with_center.svg';
-import targetWithoutCenterSvgStr from '../style/icons/target_without_center.svg';
-import terminalToolbarSvgStr from '../style/icons/terminal_toolbar.svg';
-import vectorSquareSvgStr from '../style/icons/vector_square.svg';
-import visibilitySvgStr from '../style/icons/visibility.svg';
+import bookOpenSvgStr from '../../style/icons/book_open.svg';
+import clockSvgStr from '../../style/icons/clock-solid.svg';
+import geoJsonSvgStr from '../../style/icons/geojson.svg';
+import geolocationSvgStr from '../../style/icons/geolocation.svg';
+import geoPackageSvgStr from '../../style/icons/geopackage.svg';
+import infoSvgStr from '../../style/icons/info-solid.svg';
+import logoSvgStr from '../../style/icons/logo.svg';
+import logoMiniSvgStr from '../../style/icons/logo_mini.svg';
+import logoMiniAlternativeSvgStr from '../../style/icons/logo_mini_alternative.svg';
+import logoMiniQGZ from '../../style/icons/logo_mini_qgz.svg';
+import markerSvgStr from '../../style/icons/marker.svg';
+import moundSvgStr from '../../style/icons/mound.svg';
+import nonVisibilitySvgStr from '../../style/icons/nonvisibility.svg';
+import rasterSvgStr from '../../style/icons/raster.svg';
+import targetWithCenterSvgStr from '../../style/icons/target_with_center.svg';
+import targetWithoutCenterSvgStr from '../../style/icons/target_without_center.svg';
+import terminalToolbarSvgStr from '../../style/icons/terminal_toolbar.svg';
+import vectorSquareSvgStr from '../../style/icons/vector_square.svg';
+import visibilitySvgStr from '../../style/icons/visibility.svg';
 
 export const logoIcon = new LabIcon({
   name: 'jupytergis::logo',
